+++ pnet-data-api-java/src/main/java/at/porscheinformatik/happyrest/RestLoggerAdapter.java
@@ -1,0 +1,13 @@
+package at.porscheinformatik.happyrest;
+
+/**
+ * Used for logging
+ *
+ * @author HAM
+ */
+public interface RestLoggerAdapter
+{
+
+    void logRequest(RestMethod method, String uri);
+
+}