+++ pnet-data-api-java/src/main/java/at/porscheinformatik/happyrest/spring/Spring4RestCallFactory.java
@@ -7,7 +7,9 @@ import org.springframework.web.client.RestTemplate;
 
 import at.porscheinformatik.happyrest.RestCall;
 import at.porscheinformatik.happyrest.RestCallFactory;
+import at.porscheinformatik.happyrest.RestLoggerAdapter;
 import at.porscheinformatik.happyrest.RestUtils;
+import at.porscheinformatik.happyrest.slf4j.Slf4jRestLoggerAdapter;
 
 /**
  * A factory for REST calls using spring
@@ -38,7 +40,7 @@ public class Spring4RestCallFactory implements RestCallFactory
             return execution.execute(request, body);
         });
 
-        DEFAULT = new Spring4RestCallFactory(REST_TEMPLATE, null);
+        DEFAULT = new Spring4RestCallFactory(REST_TEMPLATE, Slf4jRestLoggerAdapter.getDefault(), null);
     }
 
     public static Spring4RestCallFactory getDefault()
@@ -55,28 +57,33 @@ public class Spring4RestCallFactory implements RestCallFactory
                 return execution.execute(request, body);
             });
 
-            factory = new Spring4RestCallFactory(restTemplate, null);
+            factory = new Spring4RestCallFactory(restTemplate, Slf4jRestLoggerAdapter.getDefault(), null);
+
+            defaultFactory = factory;
         }
 
         return factory;
     }
 
     private final RestTemplate restTemplate;
+    private final RestLoggerAdapter loggerAdapter;
     private final ConversionService conversionService;
 
     @Autowired
-    public Spring4RestCallFactory(RestTemplate restTemplate, ConversionService conversionService)
+    public Spring4RestCallFactory(RestTemplate restTemplate, RestLoggerAdapter loggerAdapter,
+        ConversionService conversionService)
     {
         super();
 
         this.restTemplate = restTemplate;
+        this.loggerAdapter = loggerAdapter;
         this.conversionService = conversionService;
     }
 
     @Override
     public RestCall url(String url)
     {
-        return new Spring4RestCall(restTemplate, conversionService, url);
+        return new Spring4RestCall(restTemplate, loggerAdapter, conversionService, url);
     }
 
 }